Move/Rename existing tests
Moves all existing integration tests into a Legacy namespace so they can be easily grouped and excluded as getting these tests to run is never likely to happen.
