+++ tests/Legacy/PostmarkClientSuppressionsTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Postmark\Tests;
+namespace Postmark\Tests\Legacy;
 
 require_once __DIR__ . "/PostmarkClientBaseTest.php";
 
@@ -12,7 +12,8 @@ class PostmarkClientSuppressionsTest extends PostmarkClientBaseTest {
 
     private $testServerToken = "";
 
-    public static function tearDownAfterClass() {
+    public static function tearDownAfterClass(): void
+    {
         $tk = parent::$testKeys;
         $client = new PostmarkClient($tk->WRITE_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
     }
